backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
mistralai>=1.10.0

sentence-transformers
scikit-learn
numpy
requests
